Pull out Executor logic
